moved all editor scripts into their own Editor folders, as they need to be excluded when building for android
